refactor: CafeActionBar의 구현을 CafeCard로 이동
